不要な VStack 削除
